doc: implement openapi documentation + validation

+++ services/backend/factory-service/src/main/java/com/veteroch4k/factory_service/models/commands/MaterialReservationCommand.java
@@ -2,6 +2,6 @@ package com.veteroch4k.factory_service.models.commands;
 
 import java.util.List;
 
-public record MaterialReservationCommand(int orderId, List<RequiredMaterial> materials, int factoryId) {
+public record MaterialReservationCommand(Long orderId, List<RequiredMaterial> materials, Long factoryId) {
 
 }
