+++ services/backend/factory-service/src/main/java/com/veteroch4k/factory_service/models/commands/RequiredMaterial.java
@@ -1,6 +1,6 @@
 package com.veteroch4k.factory_service.models.commands;
 
-public record RequiredMaterial(int materialId,
-                               int quantity) {
+public record RequiredMaterial(Long materialId,
+                               Long quantity) {
 
 }
