+++ services/backend/factory-service/src/main/java/com/veteroch4k/factory_service/models/commands/ToolReservationCommand.java
@@ -3,6 +3,6 @@ package com.veteroch4k.factory_service.models.commands;
 import com.veteroch4k.factory_service.models.OperationTools;
 import java.util.List;
 
-public record ToolReservationCommand(int orderId, List<RequiredTools> tools, int factoryId) {
+public record ToolReservationCommand(Long orderId, List<RequiredTools> tools, Long factoryId) {
 
 }
