+++ services/backend/factory-service/src/main/java/com/veteroch4k/factory_service/models/events/OrderCreatedEvent.java
@@ -2,7 +2,7 @@ package com.veteroch4k.factory_service.models.events;
 
 
 public record OrderCreatedEvent(
-    int orderId,
-    int productId,
-    int productQuantity
+        Long orderId,
+        Long productId,
+        Long productQuantity
 ) {}
